+++ packages/openai/src/openai-chat-language-model.test.ts
@@ -175,6 +175,8 @@ describe('doGenerate', () => {
       completion_tokens?: number;
       completion_tokens_details?: {
         reasoning_tokens?: number;
+        accepted_prediction_tokens?: number;
+        rejected_prediction_tokens?: number;
       };
       prompt_tokens_details?: {
         cached_tokens?: number;
@@ -1088,6 +1090,35 @@ describe('doGenerate', () => {
     });
   });
 
+  it('should return accepted_prediction_tokens and rejected_prediction_tokens in completion_details_tokens', async () => {
+    prepareJsonResponse({
+      usage: {
+        prompt_tokens: 15,
+        completion_tokens: 20,
+        total_tokens: 35,
+        completion_tokens_details: {
+          accepted_prediction_tokens: 123,
+          rejected_prediction_tokens: 456,
+        },
+      },
+    });
+
+    const model = provider.chat('gpt-4o-mini');
+
+    const result = await model.doGenerate({
+      inputFormat: 'prompt',
+      mode: { type: 'regular' },
+      prompt: TEST_PROMPT,
+    });
+
+    expect(result.providerMetadata).toStrictEqual({
+      openai: {
+        acceptedPredictionTokens: 123,
+        rejectedPredictionTokens: 456,
+      },
+    });
+  });
+
   describe('reasoning models', () => {
     it('should clear out temperature, top_p, frequency_penalty, presence_penalty', async () => {
       prepareJsonResponse();
@@ -1262,7 +1293,9 @@ describe('doStream', () => {
         cached_tokens?: number;
       };
       completion_tokens_details?: {
-        reasoning_tokens: number;
+        reasoning_tokens?: number;
+        accepted_prediction_tokens?: number;
+        rejected_prediction_tokens?: number;
       };
     };
     logprobs?: {
@@ -1333,6 +1366,7 @@ describe('doStream', () => {
         finishReason: 'stop',
         logprobs: mapOpenAIChatLogProbsOutput(TEST_LOGPROBS),
         usage: { promptTokens: 17, completionTokens: 227 },
+        providerMetadata: { openai: {} },
       },
     ]);
   });
@@ -1460,6 +1494,7 @@ describe('doStream', () => {
         finishReason: 'tool-calls',
         logprobs: undefined,
         usage: { promptTokens: 53, completionTokens: 17 },
+        providerMetadata: { openai: {} },
       },
     ]);
   });
@@ -1594,6 +1629,7 @@ describe('doStream', () => {
         finishReason: 'tool-calls',
         logprobs: undefined,
         usage: { promptTokens: 53, completionTokens: 17 },
+        providerMetadata: { openai: {} },
       },
     ]);
   });
@@ -1717,6 +1753,7 @@ describe('doStream', () => {
         finishReason: 'tool-calls',
         logprobs: undefined,
         usage: { promptTokens: 226, completionTokens: 20 },
+        providerMetadata: { openai: {} },
       },
     ]);
   });
@@ -1781,6 +1818,7 @@ describe('doStream', () => {
         finishReason: 'tool-calls',
         logprobs: undefined,
         usage: { promptTokens: 53, completionTokens: 17 },
+        providerMetadata: { openai: {} },
       },
     ]);
   });
@@ -1857,6 +1895,7 @@ describe('doStream', () => {
         finishReason: 'stop',
         logprobs: undefined,
         usage: { promptTokens: 53, completionTokens: 17 },
+        providerMetadata: { openai: {} },
       },
     ]);
   });
@@ -1895,6 +1934,7 @@ describe('doStream', () => {
           completionTokens: NaN,
           promptTokens: NaN,
         },
+        providerMetadata: { openai: {} },
       },
     ]);
   });
@@ -1920,6 +1960,7 @@ describe('doStream', () => {
         completionTokens: NaN,
         promptTokens: NaN,
       },
+      providerMetadata: { openai: {} },
     });
   });
 
@@ -2011,7 +2052,7 @@ describe('doStream', () => {
     });
   });
 
-  it('should handle cached tokens in experimental_providerMetadata', async () => {
+  it('should return cached tokens in providerMetadata', async () => {
     prepareStreamResponse({
       content: [],
       usage: {
@@ -2037,10 +2078,7 @@ describe('doStream', () => {
       messages: [{ role: 'user', content: 'Hello' }],
     });
 
-    const chunksArr = await convertReadableStreamToArray(stream);
-    expect(chunksArr[chunksArr.length - 1]).toHaveProperty('providerMetadata');
-    expect(chunksArr[chunksArr.length - 1].type).toEqual('finish');
-    expect(chunksArr[chunksArr.length - 1]).toStrictEqual({
+    expect((await convertReadableStreamToArray(stream)).at(-1)).toStrictEqual({
       type: 'finish',
       finishReason: 'stop',
       logprobs: undefined,
@@ -2050,6 +2088,50 @@ describe('doStream', () => {
       },
       providerMetadata: {
         openai: { cachedPromptTokens: 1152 },
+      },
+    });
+  });
+
+  it('should return accepted_prediction_tokens and rejected_prediction_tokens in providerMetadata', async () => {
+    prepareStreamResponse({
+      content: [],
+      usage: {
+        prompt_tokens: 15,
+        completion_tokens: 20,
+        total_tokens: 35,
+        completion_tokens_details: {
+          accepted_prediction_tokens: 123,
+          rejected_prediction_tokens: 456,
+        },
+      },
+    });
+
+    const { stream } = await model.doStream({
+      inputFormat: 'prompt',
+      mode: { type: 'regular' },
+      prompt: TEST_PROMPT,
+    });
+
+    expect(await server.getRequestBodyJson()).toStrictEqual({
+      stream: true,
+      stream_options: { include_usage: true },
+      model: 'gpt-3.5-turbo',
+      messages: [{ role: 'user', content: 'Hello' }],
+    });
+
+    expect((await convertReadableStreamToArray(stream)).at(-1)).toStrictEqual({
+      type: 'finish',
+      finishReason: 'stop',
+      logprobs: undefined,
+      usage: {
+        promptTokens: 15,
+        completionTokens: 20,
+      },
+      providerMetadata: {
+        openai: {
+          acceptedPredictionTokens: 123,
+          rejectedPredictionTokens: 456,
+        },
       },
     });
   });
@@ -2133,6 +2215,7 @@ describe('doStream', () => {
           finishReason: 'stop',
           usage: { promptTokens: 17, completionTokens: 227 },
           logprobs: undefined,
+          providerMetadata: { openai: {} },
         },
       ]);
     });
@@ -2292,7 +2375,7 @@ describe('doStream simulated streaming', () => {
         finishReason: 'stop',
         usage: { promptTokens: 4, completionTokens: 30 },
         logprobs: undefined,
-        providerMetadata: undefined,
+        providerMetadata: { openai: {} },
       },
     ]);
   });
@@ -2356,7 +2439,7 @@ describe('doStream simulated streaming', () => {
         finishReason: 'stop',
         usage: { promptTokens: 4, completionTokens: 30 },
         logprobs: undefined,
-        providerMetadata: undefined,
+        providerMetadata: { openai: {} },
       },
     ]);
   });