+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,27 @@
 
 ### Patch Changes
 
+- Updated dependencies [a8f3242]
+  - ai@4.0.26
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [0823899]
+  - ai@4.0.25
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [ae0485b]
+  - ai@4.0.24
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [bc4cd19]
   - ai@4.0.23