+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/azure
 
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [b19aa82]
+  - @ai-sdk/openai@1.0.13
+
 ## 1.0.14
 
 ### Patch Changes